CalcFeedsDelta handles changed feeds correctly
With this new code that detects changed feeds, CalcFeedsDelta will now
return the changed feed in both the remove and add outputs.

The case found by Traun Leyden is when shrinking a cluster.  In that
case, pindexes are reassigned to nodes that remain in the cluster, and
the previous code was incorrectly not detecting the change and not
restarting existing feeds.