MB-17147 - optional async loading of data dir
The default behavior of synchronous loading of pindexes from the data
dir during manager.Start() remains, but there's now an option to allow
the app to specify asynchronous data dir loading (by the janitor).

This can lead to faster time-to-availability as the app can perhaps
provide an http rest endpoint faster with async data dir loading.

Change-Id: Ibab45b3ec49a724b0f1a8449beebcf3e33953fbb
Reviewed-on: http://review.couchbase.org/58022
Reviewed-by: Marty Schoch <marty.schoch@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>