MB-17541 - BREAKING API - IndexDef/PlanPIndex JSON naturalized
IndexDef and PlanPIndex structs are now marshaled to JSON in natural,
nested object format.

In contrast, the previous format treated the Params, IndexParams &
SourceParams fields as hard-to-use string'ified JSON (JSON encoded as
a string).

That is, we're now using type map[string]interface{} instead of type
string for the Params, IndexParams, SourceParams fields.

The unmarshal/parsing of those fields, though, accepts both the new,
nested object format and the old, enveloped string format for
backwards compatibility.

Change-Id: I1ea878f7ce1468c2ece721e678ecd35cf7a4de5f
Reviewed-on: http://review.couchbase.org/59297
Reviewed-by: Marty Schoch <marty.schoch@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>