MB-16794 - CBAuth() helper enables working GatherStats()
cbauth provides a cbauth.NewAuthHandler() API that does all the heavy
lifting, allowing this change to remove complex, edge-case riddled
codepaths from cbgt.

Instead, there's now a simple cbgt.CBAuth() helper func that
consolidates the code related to initializing a couchbase.AuthHandler,
such as relying on an AuthHandler that comes from cbauth, if
available.

Related, this allows the couchbase.GatherStats() API to succeed
because of the proper cluster-aware auth.

Change-Id: Ia2eed750dda0f91c3c2ab7bf46dd1ada93518a42
Reviewed-on: http://review.couchbase.org/60365
Tested-by: Steve Yen <steve.yen@gmail.com>
Reviewed-by: Marty Schoch <marty.schoch@gmail.com>