MB-18294 - CfgMetaKV nodeDef updated only by owning node
This change is regarding CfgMetaKV...

Previously, each node would write all the node-defs to the "directory"
of node-def entries in metakv, and that led to racing nodes to stomp
on each other's work.  Especially a slower node would incorrectly
overwrite the node-def entry of a node that was trying to publish its
latest HostPort assignment.

With this change, each node only write its own node-def entry.

Change-Id: I8766963bcf46d56a0d7ac0544a2b228e8724a776
Reviewed-on: http://review.couchbase.org/60406
Tested-by: Steve Yen <steve.yen@gmail.com>
Reviewed-by: Marty Schoch <marty.schoch@gmail.com>