MB-18385, MB-18888 - MCP caught up to primary pindex not feed
In this change, the MCP considers a pindex "move" to be "caught up"
when the moved pindex has reached the same seq-num as the (former)
primary pindex, if it exists.

Previously, the implementation would treat the source partition's
seq-num as the basis for the "caught up or not" comparison.

Also, as part of MB-18888, with this change, only primary pindexes
will go through the caught-up checking, and replica pindexes will not.

Change-Id: I29613cca570774c0437d38fe02c861466b4414a9
Reviewed-on: http://review.couchbase.org/61983
Reviewed-by: Marty Schoch <marty.schoch@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>