MB-19117 - DeleteAllIndexFromSource finishes loop, even on err
The theory with MB-19117 / DeleteAllIndexFromSource is that it runs
into some race condition (perhaps CAS mismatch?) where it errors out
early in its loop through all the indexDefs.

With this change, it remembers the first error, but tries to complete
the loop in any case, trying to be more bulletproof.

Change-Id: Iaf046047ea19b61e3d39efaf7313b48b27ba7dcf
Reviewed-on: http://review.couchbase.org/62593
Reviewed-by: Marty Schoch <marty.schoch@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>