MB-19207 - non-blocking progress updates to fix stop-rebalance
Before this change, the behavior was that during a Stop Rebalance
operation, the rebalancer would acquire a lock and then proceed with
rebalance cancellation activities.  Unfortunately, as part of that,
there were progress updates, and the progress updater code would want
to acquire the same lock.

This change breaks the deadlock by refactoring the progress updater
code onto its own goroutine with a buffered channel in front of it, so
that updates can happen asynchronously.

Change-Id: I3436308b5f9aed5a2d2fc1791e3e59eb2508e0cc
Reviewed-on: http://review.couchbase.org/62862
Tested-by: Steve Yen <steve.yen@gmail.com>
Reviewed-by: Marty Schoch <marty.schoch@gmail.com>
Reviewed-by: Steve Yen <steve.yen@gmail.com>