MB-20050 - refresh mgr.lastIndexDefs cache on DDL
I don't see this issue running the unit tests on my dev box, but the
CI/jenkins systems shows there's a race where a client that tries to
retrieve stats right after index definition might incorrectly get a
index-not-found kind of error.

This proposed fix refreshes the mgr.lastIndexDefs cache as part of the
index definition handling to try to close the race.

Change-Id: I0276de23a598f9ff20b6583494f573264ed83daa
Reviewed-on: http://review.couchbase.org/65779
Reviewed-by: Marty Schoch <marty.schoch@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>