optional PlannerHooks for advanced planning overrides
Some advanced users of cbgt may need to override the default pindex
planning algorithm at certain points.  This commit introduces an
optional PlannerHooks API so that the application can provide
callbacks at various points in the planning cycle.

An example might be applications that have indexes that have large,
single shards (e.g., just a single pindex per indexDef).  An advanced
application might want to dynamically adjust the node weights during
the planning algorithm in order to reach an improved balancing, and
can do so via a PlannerHook.

As part of this change, indexDef's are now processed in order of their
sorted name, in contrast to the previous, undefined hashmap iteration
order.

Change-Id: Ia0eb3990d891c0f400958ba8a97cc871b30d5f58
Reviewed-on: http://review.couchbase.org/67562
Reviewed-by: Marty Schoch <marty.schoch@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>