MB-20205 - cached nodeDefs when calculating covering pindexes
To process a query, a set of covering pindexes (or shards) for an
index need to be computed.

As part of that, the previous code would lookup the latest nodeDefs in
the Cfg.  When the Cfg is metakv, that would mean a REST call from
cbft to ns-server which may be slow.  Additionally, that Cfg retrieval
would happen while holding the cbgt manager's lock.  Quite slow.

For some reason, I had mistakenly thought cbauth/metakv provided
caching, but it apparently does not.

Props to Alex Gyryk for finding the clues here, noticing that
ns-server CPU was extremely high during perf tests, and from Pavel
Paulau for the hint to check ns-server's access log.

This change introduces cached nodeDefs, which has the downside that
under a topology change, the system might have a longer window of
attempting to use an outdated, previously cached nodeDefs, meaning
that queries may see more partial result errors during rebalance.

Change-Id: Icf13a99adf44c2025a762fec8fd0146ed6f82656
Reviewed-on: http://review.couchbase.org/67975
Reviewed-by: Marty Schoch <marty.schoch@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>