MB-21219 - fix covering pindexes cache invalidation
Issue MB-21219 was introduced with the performance optimization that
cached the covering pindexes computation, that helped to speed up
queries.  But, the cache wasn't correctly invalidated in all cases,
especially in a multi-node cluster situation.

In this commit, when the pindex instances that are local to a node are
changed (are registered/unregistered), the covering pindex cache
should be invalidated.

Change-Id: Idb926118b94489f5aa427e6f245a574cfce1e701
Reviewed-on: http://review.couchbase.org/68341
Reviewed-by: Steve Yen <steve.yen@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>