MB-17433 - store smaller planPIndexes into cfgMetaKv
This change is based on the observation that many planPIndexes stored
into the Cfg have shared index definitions and source definitions.

Instead of repeating those definitions for every pindex (e.g.,
repeated 32 times for 32 pindexes), the idea is to store the
information once, and reconstitute it during a Cfg.Get().

Change-Id: I76debb63dc8fe74f1d6c94b35f44b21271d27551
Reviewed-on: http://review.couchbase.org/68906
Reviewed-by: Marty Schoch <marty.schoch@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>