MB-23257 - respect defaults in IndexDef custom UnmarshalJSON()
Looks like this bug came in awhile ago with the "fix" to handle the
so-called nested vs enveloped IndexDef JSON formats.

In the custom IndexDef.UnmarshalJSON() method, the fix here is that we
need to keep and respect any default field values that the caller
might have setup already in the IndexDef, instead of starting from a
zero/empty IndexDef.

Change-Id: I8cdd5c41c79dfc6f572679ef2808f8a62c0abad0
Reviewed-on: http://review.couchbase.org/75240
Reviewed-by: Marty Schoch <marty.schoch@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>