MB-24249: Avoid handlers writing http response headers twice
When encoding/json's NewEncoder()'s Encode is invoked, for smaller
responses the API succeeds and does not return a non-nil error.
However in case of larger responses, an error is seen sometimes:

    "write tcp 127.0.0.1:9200->127.0.0.1:49239: write: broken pipe"

This is because large responses are interrupted by a timeout/deadline
set on the connection by the client, or in other cases the socket being
simply closed by the client.

Sequence of steps to produce the message "http: multiple
response.WriteHeader calls" in the logs:
1. Encode() invokes the responseWriter's Write. This Write first
   invokes WriteHeader() with code 200 (per go docs) before writing
   the data on the stream. However while progress ensues in writing
   the large response, the client closes the connection which interrupts
   the server.
2. Encode() therefore returns a non-nil error because of the client
   interruption, and thus invokes WriteHeader() again - this time with
   the code 500, causing the log message in focus.

Since this is an error we can actually ignore, for the client has
already closed the connection and doesn't really care - I recommend we
stick to the earlier cover-up fix made here:
    http://review.couchbase.org/#/c/79272/

In the current commit, I'm just changing the qualifier of 'wrote' in
CountResponseWriter to public, so that it can be accessed in CBFT
where applicable.

Change-Id: Ie392374d633e6206615012d4770e31be0ba026fa
Reviewed-on: http://review.couchbase.org/88639
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Steve Yen <steve.yen@gmail.com>
Reviewed-by: Marty Schoch <marty.schoch@gmail.com>
Tested-by: Abhinav Dangeti <abhinav@couchbase.com>