MB-31965 - converingCache unused
The CAS like version isn't really used in the coveringCache
implementation here. The current incorrect use of version is
blocking the initialisation of coveringCache and it remains
un used. As the cache gets cleaned with any change to any of
those arguments which affect the covering pindexes computation,
the need of ver looks redundant.

Change-Id: Ie361f12082a66d3b533231896c2d1b2820fcb12f
Reviewed-on: http://review.couchbase.org/101631
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Steve Yen <steve.yen@gmail.com>
Tested-by: Sreekanth Sivasankaran <sreekanth.sivasankaran@couchbase.com>