MB-35508 - Index wiped off when running plan freeze
During the newer plan computations in the planner,
the index UUID checks were leaving the new plans empty and
hence this wipe out was happening.  As the index definition
update is happening during a plan Freeze, the UUID bump was
expected.
But it is still safe to copy over the older pindex
contents with an extra check for confirming the nature
of index definitions changes.

The scope of this change is still very much restricted as
this is applicable only when the plan is frozen.

Change-Id: Icbaa24eb66a610c23e9a953b19bbbb352d9bcdb5
Reviewed-on: http://review.couchbase.org/113112
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Abhinav Dangeti <abhinav@couchbase.com>
Tested-by: Sreekanth Sivasankaran <sreekanth.sivasankaran@couchbase.com>