MB-44703 - fts swap rebalance reports progress > 100%
This happens due to incorrect calculation of the number
of partitions getting moved during a rebalance. And that
stems from the lack of knowledge about the number of nodes
in a pre-rebalance cluster. This could be fixed by
deferring the self-registeration of a node to Cfg during
boot up.
This fix attempts to improve the logs during the ctl startup.

Change-Id: Ie98d186f9cd5900f73352288a9d1548d196e63a1
Reviewed-on: http://review.couchbase.org/c/cbgt/+/148020
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Abhinav Dangeti <abhinav@couchbase.com>
Tested-by: Sreekanth Sivasankaran <sreekanth.sivasankaran@couchbase.com>