MB-45169 - FTS Failover exited with reason {{timeout}}
StartTopologyChange could take a lot of time due to the
time consuming calls to the feedType.Partitions() for all
the indexes. This could be significant with a high number
of fts indexes.
Many of these calls could be saved for reuse as the only
variable arguments for partition computations  are the
sourceName and the sourceUUID.

Change-Id: Ic0bad474ee2e7f75b678872e3b7ea515e5e0c214
Reviewed-on: http://review.couchbase.org/c/cbgt/+/149602
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Sreekanth Sivasankaran <sreekanth.sivasankaran@couchbase.com>
Reviewed-by: Sreekanth Sivasankaran <sreekanth.sivasankaran@couchbase.com>