MB-44446 remove Manager Lock/Unlock
Having exported Lock/Unlock methods makes it difficult to track
who is obtaining a lock and for what purpose.  This should
remain an internal implementation detail.

Here we refactor the one method that required it, into a new
method which keeps the lock internal to the Manager, while still
supporting the same functionality.

Change-Id: Ie2d4f625efe19d7b28863a431e7f9035049a5f3a
Reviewed-on: http://review.couchbase.org/c/cbgt/+/149785
Tested-by: Marty Schoch <marty.schoch@gmail.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Abhinav Dangeti <abhinav@couchbase.com>
Reviewed-by: Sreekanth Sivasankaran <sreekanth.sivasankaran@couchbase.com>