Incorrect usage of errors.Unwrap
errors.Unwrap(errNoRetryNeeded) returns a nil as errNoRetryNeeded
does not implement an Unwrap() method.
See: https://golang.org/pkg/errors/#Unwrap

If the intention here was to stop re-trying and return an error
when retrying is to NOT be attempted, then this change does
exactly that.

Change-Id: I5e3b55d49298ee78d39b07fad772cc978155840d
Reviewed-on: http://review.couchbase.org/c/cbgt/+/150305
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Marty Schoch <marty.schoch@gmail.com>
Tested-by: Abhinav Dangeti <abhinav@couchbase.com>