MB-46312: Rebalance progress is stuck indefinitely
With this change, the rebalance progress monitor considers
the minimum of the below values for assessing the progress
completion.

-Partition sequence numbers from the former primary node.
-Minimum value of the source sequence number stat between
the new node and former primary node for the partition.

[Source partition seq numbers should be same across
former/new fts nodes, but this is to be on the safer
side with any unforeseen kv rebalance/failover
complications]

Change-Id: I67c4a91db2607d1f63b52ae72c6cdeb3c42f6819
Reviewed-on: http://review.couchbase.org/c/cbgt/+/154006
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Abhinav Dangeti <abhinav@couchbase.com>
Tested-by: Abhinav Dangeti <abhinav@couchbase.com>