MB-48514 - Rebalance taking longer with high #indexes
The control handshakes between various blance's
orchestrator routines(mover and supply) is seen
to be slow with higher number of indexes in a
busy cluster.
The fix here is to skip the orchestrator invocations
if there is no plan change(node reassignments) for
a given index during the rebalance.

Change-Id: I91252309554b8a11ea887be28b80403032939854
Reviewed-on: http://review.couchbase.org/c/cbgt/+/162989
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Abhinav Dangeti <abhinav@couchbase.com>
Tested-by: Sreekanth Sivasankaran <sreekanth.sivasankaran@couchbase.com>