MB-47351 - FTS scorch AsyncError, treating this as fatal
The root cause for this error is the concurrent boot up
attempt of a partition by both the manager and the janior
threads during the initial boot up of the process.
The fix attempts to complete the booting status updates
from the Janitor's side as well. And that should prevent
the concurrent/racy attempts here.

Change-Id: Ibe4d3d8ee9ee0579b1ba13d420d69f6ecab7edc2
Reviewed-on: http://review.couchbase.org/c/cbgt/+/163365
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Abhinav Dangeti <abhinav@couchbase.com>
Tested-by: Sreekanth Sivasankaran <sreekanth.sivasankaran@couchbase.com>