MB-50347 - Failover/rebalance stuck for more than 18 hrs
Adding UUID verifications during the plan update operations
so that it can retry the planning operations if there was a
change in the metakv plan from a concurrent rebalance
planner in the background. This ought to help the index
Create/Update/Delete handling during a rebalance.

Change-Id: I41e4a6e0bc2b70c71dba2ded8f2d3be930ef47ae
Reviewed-on: https://review.couchbase.org/c/cbgt/+/170441
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Abhinav Dangeti <abhinav@couchbase.com>
Reviewed-by: <thejas.orkombu@couchbase.com>
Tested-by: Sreekanth Sivasankaran <sreekanth.sivasankaran@couchbase.com>