MB-53348 - 1 partition/1 replica index gets rebuild over failover-recovery
The stable plan used for restoring the cluster state to the pre-failover
scenario was getting overridden by failed-over plan. And this was
resulting in the partition-node reassignments.
Updated the stable plan computations to accommodate the total partition
count as well while verifying for a stable/complete plan for the index.

Change-Id: Idc1a36708237a20018ca95a41fa719142c89a4b0
Reviewed-on: https://review.couchbase.org/c/cbgt/+/178749
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Abhi Dangeti <abhinav@couchbase.com>
Reviewed-by: Abhi Dangeti <abhinav@couchbase.com>