MB-55895: Rebalance Fixes
1. Currently, rebalance uses plans fetched before the
first iteration, which are passed to blance.
These may not be reflective of the latest plans(which
may have been changed by blance).

a. This PR fetches plans from Cfg each time they
need to be sent to blance.
b. This PR also adds a field which reflects these
latest plans and is used by the serverless rebalance
hook.

2. Currently, auto rebalance has a non-empty nodesToAdd
when some of the nodes do not contain partitions. The
expected nodesToAdd here is an empty list.

This PR compares the member nodes with the existing
nodes to determine if any nodes are to be added.

Change-Id: I311cb550bd74686c8bcef67837f38bde3e64f6b4
Reviewed-on: https://review.couchbase.org/c/cbgt/+/187809
Tested-by: Aditi Ahuja <aditi.ahuja@couchbase.com>
Reviewed-by: Abhi Dangeti <abhinav@couchbase.com>