MB-57386: Avoid using task orchestrator for pause and resume
The orchestrator field being set is used as a quick check
for rebalance status before querying an endpoint.

Currently, since pause/resume share a common orchestrator field
with rebalance, there are times when a rebalance is falsely
detected by ctl on the orchestrator node. This prevents the
resume from proceeding.

This PR avoids the use of a common variable to fix this.

Change-Id: I8e425dbe1dbea2f7dbf2e7718cf1d05fdb188549
Reviewed-on: https://review.couchbase.org/c/cbgt/+/191655
Tested-by: Aditi Ahuja <aditi.ahuja@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Abhi Dangeti <abhinav@couchbase.com>