MB-57396: Directory creation when creating partitions
There are common steps to creating new partitions and
rolling back a partition.

This PR creates a directory before storing
pindex meta file for cases such as full rollback
of the pindex, due to which the path may not
have been created before saving the PINDEX_META
file.

This PR also avoids a panic in partial rollback
when there is an error opening a pindex by
storing the path.

Change-Id: I6f4a9015f4c3c5cea23a42dbff851ab2085459a4
Reviewed-on: https://review.couchbase.org/c/cbgt/+/193034
Well-Formed: Restriction Checker
Reviewed-by: Abhi Dangeti <abhinav@couchbase.com>
Tested-by: Aditi Ahuja <aditi.ahuja@couchbase.com>