MB-57334: [BP] Fixing nodes to add list
Currently, auto rebalance has a non-empty nodesToAdd
when some of the nodes do not contain partitions. The
expected nodesToAdd here is an empty list.

This PR adds a new function which compares the member
nodes with the existing nodes to determine if any nodes
are to be added.

Change-Id: I6e17e0dbcf331a2bd3bddc6032c583008c965251
Reviewed-on: https://review.couchbase.org/c/cbgt/+/193567
Well-Formed: Restriction Checker
Reviewed-by: Mohd Shaad Khan <mohd.shaadkhan@couchbase.com>
Tested-by: Aditi Ahuja <aditi.ahuja@couchbase.com>