MB-59024: Compute plans after stopped rebalance
Currently, if an index is introduced during a
rebalance and the rebalance is cancelled before
the index is rebalanced, no plans are computed
for the index and consequently, no partitions
are created.

This PR fixes that by computing plans at the
end of a rebalance, whether stopped or completed.

Change-Id: Ic1f36dc3a0e8ba266b0131a9c9db539e1bf16cdd
Reviewed-on: https://review.couchbase.org/c/cbgt/+/199239
Well-Formed: Build Bot <build@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Abhi Dangeti <abhinav@couchbase.com>
Tested-by: Aditi Ahuja <aditi.ahuja@couchbase.com>