MB-59409: Honoring the indexPartitions setting within definitions
A bug in the conversion of indexPartitions into maxPartitionsPerPIndex
can cause the planner to go off by a bit while actually generating
the index plans.
NOTE that this issue does NOT cause any form of data loss but instead -
sometimes can cause more or less number of index partitions than
actually expected.

For example, if indexPartitions for a bucket with 1024 vbuckets is
chosen to be 60, then maxPartitionPerPIndex is set to 18, which means
we actually end up generating 57 index partitions that take on no
more than 18 vbuckets each.
    => 56 partitions w/ 18 vbs each + 1 partition w/ 16 vbs

The fix ensures that in the above scenario there, the outcome is
60 partitions with this distribution ..
    => 56 partitions w/ 17 vbs each + 4 partitions w/ 18 vbs each

Change-Id: I206afed672ce5c3d8282d05896c4789a13f0c69a
Reviewed-on: https://review.couchbase.org/c/cbgt/+/200005
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: <thejas.orkombu@couchbase.com>
Tested-by: Abhi Dangeti <abhinav@couchbase.com>