MB-61043: Fix computation of nodes to remove
Currently, on a rebalance-in following a
rebalance-out, the nodes to remove list
is incorrectly populated when it should
be empty. This incorrectly identifies
the rebalance as a swap rebalance.

This is due to populating the Ctl's member
nodes based on the node defs in cfg, which
may include the node being removed.

This PR
- uses the topology change request sent by
ns_server to determine the expected nodes
after the rebalance and uses those to
populate ctl's member nodes during the Prepare
phase.

- refactors functions called after a Prepare
task into rebalance and non-rebalance functions.

(Refer to MB-61185 for more).

Change-Id: Ia630991d68330c571cf60999ba5b836554af070f
Reviewed-on: https://review.couchbase.org/c/cbgt/+/207290
Well-Formed: Restriction Checker
Reviewed-by: Abhi Dangeti <abhinav@couchbase.com>
Tested-by: Abhi Dangeti <abhinav@couchbase.com>