MB-61310 - Introducing an option for catchup on replica partitions
Enabling the replica catchup would help in situations where the file
based rebalance for a replica partition fell back to the old DCP way.

During this index build if the node having the active partition
got failed over, then the building replica would get promoted and in
the rebalance of that topology change if another node got assigned
the replica of the partition then it would stream from the building
(now) active partition. Attaching feeds in this case lead to rollbacks
due to purge seq no. check.

Hence to keep the partitions in a more consistent state post a rebalance
operation, having the replicas catchup as well would be necessary.

Change-Id: I7fa7b75d535c7c06ca9c31c9af56f8b57c47d6ef
Reviewed-on: https://review.couchbase.org/c/cbgt/+/207808
Well-Formed: Restriction Checker
Reviewed-by: Abhi Dangeti <abhinav@couchbase.com>
Tested-by: <thejas.orkombu@couchbase.com>