MB-61043 - Ensure sorted order for index rebalance
Given the same inputs(existing partition layout,
node layout), the node partition assignment is
idempotent.
Earlier, since the indexes were not rebalanced
in the same order each time, resulting in
scrambling index partitions even when balanced.
This lead to the addition of a check to early
exit rebalance if there was no topology change.

Instead of the early exit check, this PR
proposes to sort index names prior to rebalance
so that even a rebalance without topology change
will not result in moving partitions.

(Pls refer to MB-61317 as well).

Change-Id: I49d197795a41f5fceae44478fa35ac11cad16e59
Reviewed-on: https://review.couchbase.org/c/cbgt/+/207916
Well-Formed: Restriction Checker
Reviewed-by: Abhi Dangeti <abhinav@couchbase.com>
Tested-by: Abhi Dangeti <abhinav@couchbase.com>