MB-63049: Introducing Planner Version
This PR introduces the notion of a planner
version to account for differences in how
the plan algorithm computes plans.
These differences can result in extra partitions
when upgrading from nodes with an older planner
to a newer one.

To rectify this:
- Avoid using older plans for provisioning and
during recovery/non-recovery rebalances.
- Introduce new version var for planner algo.
- Add a field PlannerVersion for each plan pindex
that's updated when the plan is computed.

Since the PlannerVersion field being introduced will
also be persisted, VERSION is also bumped.
If only the planner algorithm is being changed, this
doesn't need to be bumped.

Change-Id: Iaa362dbe1bea730830929ffdd38a2e912ed019a3
Reviewed-on: https://review.couchbase.org/c/cbgt/+/214041
Reviewed-by: Abhi Dangeti <abhinav@couchbase.com>
Tested-by: Aditi Ahuja <aditi.ahuja@couchbase.com>
Reviewed-by: Mohd Shaad Khan <mohd.shaadkhan@couchbase.com>
Well-Formed: Restriction Checker
Well-Formed: Build Bot <build@couchbase.com>