MB-64885: Backport 2 over v1.3.9 for SG
MB-64288 - Fix partial rollback error handling

In the current error handling for partial rollback,
the pindex is opened irrespective of a successful or
failed partial rollback but is not guaranteed to be
unregistered before the full rollback.

This PR removes the condition of an exact pointer
match for the pindex being rolled back and
unregisters it based on name.

Change-Id: Ic6caeb69d26d33a9627aee624b38b40dc450a7e3
Reviewed-on: https://review.couchbase.org/c/cbgt/+/219890
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Aditi Ahuja <aditi.ahuja@couchbase.com>
Reviewed-by: Abhi Dangeti <abhinav@couchbase.com>
Reviewed-by: Rahul Rampure <rahul.rampure@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-on: https://review.couchbase.org/c/cbgt/+/222532
Tested-by: Abhi Dangeti <abhinav@couchbase.com>