MB-58901: update QueryCtl for global scoring
Adds a new boolean parameter "GlobalScoring" to QueryCtl which
decides whether to perform the scoring of doc hits for a query
at a global(cluster) level or not.

This would be useful to keep the scoring consistent (roughly)
across varying partition counts. This might also come handy in
cases where there is a heavy skew in data distribution across the partitions.

Change-Id: I597d78db40b9bab63a61c606c737dde99bdab874
Reviewed-on: https://review.couchbase.org/c/cbgt/+/221670
Reviewed-by: Abhi Dangeti <abhinav@couchbase.com>
Tested-by: <thejas.orkombu@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>